Wait for REST layer before version check
With this commit we ensure that we wait for a little while for
Elasticsearch's REST layer to be reachable before we attempt to retrieve
its distribution version via a REST API call. This can be especially
problematic with the new `start` subcommand in automated scripts: The
`start` command will only wait for the PID file to be available and will
then terminate. If we start a benchmark immediately afterwards (that's
what happens in the integration tests), then the REST API layer might
not be up by then and the connection attempt is refused resulting in an
error.
