+++ esrally/mechanic/mechanic.py
@@ -259,6 +259,8 @@ def cluster_distribution_version(cfg, client_factory=client.EsClientFactory):
     hosts = cfg.opts("client", "hosts").default
     client_options = cfg.opts("client", "options").default
     es = client_factory(hosts, client_options).create()
+    # unconditionally wait for the REST layer - if it's not up by then, we'll intentionally raise the original error
+    client.wait_for_rest_layer(es)
     return es.info()["version"]["number"]
