+++ tests/driver/driver_test.py
@@ -1012,49 +1012,3 @@ class ProfilerTests(TestCase):
         self.assertEqual(2, return_value)
         duration = end - start
         self.assertTrue(0.9 <= duration <= 1.2, "Should sleep for roughly 1 second but took [%.2f] seconds." % duration)
-
-
-class RestLayerTests(TestCase):
-    @mock.patch("elasticsearch.Elasticsearch", autospec=True)
-    def test_successfully_waits_for_rest_layer(self, es):
-        self.assertTrue(driver.wait_for_rest_layer(es, max_attempts=3))
-
-    # don't sleep in realtime
-    @mock.patch("time.sleep")
-    @mock.patch("elasticsearch.Elasticsearch", autospec=True)
-    def test_retries_on_transport_errors(self, es, sleep):
-        import elasticsearch
-
-        es.info.side_effect = [
-            elasticsearch.TransportError(503, "Service Unavailable"),
-            elasticsearch.TransportError(401, "Unauthorized"),
-            {
-                "version": {
-                    "number": "5.0.0",
-                    "build_hash": "abc123"
-                }
-            }
-        ]
-        self.assertTrue(driver.wait_for_rest_layer(es, max_attempts=3))
-
-    # don't sleep in realtime
-    @mock.patch("time.sleep")
-    @mock.patch("elasticsearch.Elasticsearch", autospec=True)
-    def test_dont_retries_eternally_on_transport_errors(self, es, sleep):
-        import elasticsearch
-
-        es.info.side_effect = elasticsearch.TransportError(401, "Unauthorized")
-        self.assertFalse(driver.wait_for_rest_layer(es, max_attempts=3))
-
-    @mock.patch("elasticsearch.Elasticsearch", autospec=True)
-    def test_ssl_error(self, es):
-        import elasticsearch
-        import urllib3.exceptions
-
-        es.info.side_effect = elasticsearch.ConnectionError("N/A",
-                                                            "[SSL: UNKNOWN_PROTOCOL] unknown protocol (_ssl.c:719)",
-                                                            urllib3.exceptions.SSLError(
-                                                                "[SSL: UNKNOWN_PROTOCOL] unknown protocol (_ssl.c:719)"))
-        with self.assertRaisesRegex(expected_exception=exceptions.SystemSetupError,
-                                    expected_regex="Could not connect to cluster via https. Is this an https endpoint?"):
-            driver.wait_for_rest_layer(es, max_attempts=3)
